ramips: mt7621-dts: we420223-99: mux phy0->gmac1
This gives each port an individual link to the CPU, like in f1c9afd801
(ramips: mt7621-dts: mux phy0/4 to gmac1, 2022-07-06).

The advantage of this is that it can now route packets faster between
the ports (before the CPU only had a single 1Gb link to the switch that
has to be shared between both ports. Another advantage is that in Linux
5.10 you can now bridge a VLAN to a non-vlan port. Without this patch,
you're not getting any data across the bridge. That is fixed in Linux
5.15 but is still handled by the CPU in any case. So therefore this
patch is advantageous in all cases except for when you need the device
as a simple switch without VLANs. For that case it's better to revert
this and the switch hardware will forward traffic without bothering
the CPU.

Signed-off-by: Harm Berntsen <git@harmberntsen.nl>